prepare the tests to share indexes with api key
